+++ homeassistant/components/jvc_projector/config_flow.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from collections.abc import Mapping
 from typing import Any
 
-from jvcprojector import JvcProjectorAuthError, JvcProjectorConnectError
+from jvcprojector import JvcProjector, JvcProjectorAuthError, JvcProjectorConnectError
 from jvcprojector.projector import DEFAULT_PORT
 import voluptuous as vol
 
@@ -15,7 +15,6 @@ from homeassistant.data_entry_flow import FlowResult
 from homeassistant.helpers.device_registry import format_mac
 from homeassistant.util.network import is_host_valid
 
-from . import get_mac_address
 from .const import DOMAIN, NAME
 
 
@@ -118,3 +117,13 @@ class JvcProjectorConfigFlow(ConfigFlow, domain=DOMAIN):
 
 class InvalidHost(Exception):
     """Error indicating invalid network host."""
+
+
+async def get_mac_address(host: str, port: int, password: str | None) -> str:
+    """Get device mac address for config flow."""
+    device = JvcProjector(host, port=port, password=password)
+    try:
+        await device.connect(True)
+    finally:
+        await device.disconnect()
+    return device.mac
